file: Add test for endcount calculation with differing sectionsize/branches
Test fails; implementation not yet updated
